Manually merged commit from jffarge: Each class has its own file. Every errors are logged with node-color-log for better readability. Added errors to autogenerated references by the docusaurus-typedoc-plugin Removed logs from sdk Added optional pretty logs to base error
