+++ sdk/nodejs/common/utils.ts
@@ -1,0 +1,4 @@
+import logger from "node-color-log"
+
+export const log = (stack?: string) =>
+  logger.bgColor("red").color("black").log(stack)
