+++ sdk/nodejs/test/connect.spec.ts
@@ -1,6 +1,6 @@
 import { connect } from "../connect.js"
 import assert, { AssertionError } from "assert"
-import { GraphQLRequestError } from "../common/errors.js"
+import { GraphQLRequestError } from "../common/errors/index.js"
 
 describe("NodeJS sdk", function () {
   it("Connect to local engine and execute a simple query to make sure it does not fail", async function () {
